requirements.txt
amqp==5.0.5
asgiref==3.3.1
astroid==2.5.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-celery-beat==2.2.0
django-extensions==3.1.2
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-api-key==2.*
drf-yasg==1.20.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.1
packaging==20.9
pandas==1.2.2
print-dict==0.1.16
prompt-toolkit==3.0.16
pyarrow==3.0.0
pylint==2.7.2
pyparsing==2.4.7
pycparser==2.20
pyreadr==0.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
zope.event==4.5.0
zope.interface==5.2.0
